add kwargs to log function

+++ espnet2/curriculum/curriculum_logger.py
@@ -6,17 +6,22 @@ class CurriculumLogger:
     """
     Simple logger class that logs necessary stats in the log_dir.
     """
-    def __init__(self, log_dir):
+    def __init__(self, log_dir, algo, restore=False):
         if not os.path.exists(log_dir):
             os.makedirs(log_dir)
         self.log_dir = log_dir
 
         self.stats_path = os.path.join(self.log_dir, "generator_stats")
         self.policy_path = os.path.join(self.log_dir, "policy")
+        if algo=='exp3s':
+            self.weights_path = os.path.join(self.log_dir, "policy_weights")
 
-        if os.path.exists(self.stats_path):
-            os.remove(self.stats_path)
-            os.remove(self.policy_path)
+        if not restore:
+            if os.path.exists(self.stats_path):
+                os.remove(self.stats_path)
+                os.remove(self.policy_path)
+                if algo=='exp3s':
+                    os.remove(self.weights_path)
 
     def log(self, 
             iepoch, 
@@ -34,14 +39,19 @@ class CurriculumLogger:
             log_wandb (bool): logging stats to wandb
             algo (str): EXP3S or UCB
         '''
-
         with open(self.stats_path, 'a+') as fo:
-            stats = ', '.join([str(iepoch), str(iiter),\
-                               str(k), str(losses[0]), \
-                               str(losses[1]), str(progress_gain), str(reward)])
-            fo.write(stats + '\n')
-        with open(self.policy_path, 'a+') as fo:
-            fo.write(str(iepoch)+', '+str(iiter)+', '+str(policy)+'\n')
+                stats = ', '.join([str(iepoch), str(iiter),\
+                                str(k), str(kwargs["losses"][0]), \
+                                str(kwargs["losses"][1]), 
+                                str(kwargs["progress_gain"]), 
+                                str(kwargs["reward"])])
+                fo.write(stats + '\n')
+            with open(self.policy_path, 'a+') as fo:
+                fo.write(str(iepoch)+', '+str(iiter)+', '+str(policy)+'\n')
+
+        if kwargs["algo"]=='exp3':
+            with open(self.weights_path, 'a+') as fo:
+                fo.write(str(iepoch)+', '+str(iiter)+', '+str(policy)+'\n')
         
         if log_wandb:
             log_dict = {"loss":losses[1],
